refactor: split items handler into submodules
Split 1301-line items.rs into items/ directory with mod.rs, validation.rs,
crud.rs, placement.rs, and calendar.rs. Replace fetch_item_ids_in_list
with shared fetch_ordered_ids from helpers, removing #[allow(dead_code)].
